==> 02_codes/03_00_analyze_predictions.ipynb <==
import os
import pandas as pd
import matplotlib.pyplot as plt

kmeans_predictions = pd.read_csv('../03_results/out_predictions/predictions_kmeans.tsv', sep='\t', index_col=0)
kmeans_predictions.head()

def parse_tuple_string(cell):
    # Remove parentheses and split by comma
    cell = cell.strip('()')
    # Split into a tuple of floats
    avg_accuracy, avg_f1 = map(float, cell.split(','))
    return avg_accuracy, avg_f1

parsed_kmeans_predictions = kmeans_predictions.map(parse_tuple_string)

f1_scores = parsed_kmeans_predictions.map(lambda x: x[1])

max_f1_values = f1_scores.max()

max_f1_cells = f1_scores.eq(max_f1_values)

for col in max_f1_cells.columns:
    max_rows = max_f1_cells.index[max_f1_cells[col]]
    for row in max_rows:
        print(f"Max F1 in column '{col}' found in row '{row}': {parsed_kmeans_predictions.loc[row, col][1]}")

f1_scores_flat = []
for row in f1_scores.index:
    for col in f1_scores.columns:
        f1_scores_flat.append((f1_scores.loc[row, col], row, col))

f1_scores_flat_sorted = sorted(f1_scores_flat, key=lambda x: x[0], reverse=True)
f1_scores_flat_sorted

n = 200

f1_scores_flat = []
for row in f1_scores.index:
    for col in f1_scores.columns:
        f1_scores_flat.append((f1_scores.loc[row, col], row, col))

f1_scores_flat_sorted = sorted(f1_scores_flat, key=lambda x: x[0], reverse=True)

top_n_f1_scores = f1_scores_flat_sorted[:n]

top_n_df = pd.DataFrame(top_n_f1_scores, columns=['F1 Score', 'Row', 'Column'])

top_n_df.to_csv(f'../03_results/out_predictions/top_{n}_best_f1.tsv', sep='\t')

valid_matrix_types = ['M1', 'guidi', 'salazar', 'stress']

def filter_rows(row):
    column_val = row['Column']
    parts = column_val.split('_')
    if parts[0] == 'clr' and parts[1] in valid_matrix_types:
        m = int(parts[-1])
        return m >= 5
    return False

filtered_top_n_df = top_n_df[top_n_df.apply(filter_rows, axis=1)]
filtered_top_n_df.to_csv(f'../03_results/out_predictions/top_{n}_best_f1_filtered.tsv', sep='\t')
